+++ src/containerapp/HISTORY.rst
@@ -4,7 +4,10 @@ Release History
 ===============
 Upcoming
 +++++++
+* 'az containerapp up': fix --location comparison logic
+* 'az containerapp update': change --max-replicas limit
 * Add CLI support for containerapp ingress sticky-sessions'
+* Change quickstart image
 
 0.3.27
 ++++++